Video viewport on index page
Currently hard-coded to cam1/feed, but I'll make it dynamic in time.

This change also includes css to wrap the img's containing div tightly
to the img. Infuriatingly, the div's computed size is 5px taller than
the img height. Adding a -5px bottom marging to the image fixes this,
thought it's a little hacky.

Also remove the default welcome element.