Terminate the image stream if the request is disconnected
Basically we want to stop streaming the image feed if the client closes
the browser window. Connect the request "end" and "close" signals to a
method that (tries really hard to) terminate the stream.

Although this seems to work, I'm not sure which of the statements in
closeStream() does the trick, and note that it can still take some time
(many seconds) for the stream to actually stop streaming data.